Text associated with objects in the armoury
This text is to displayed on a left click on objects in the armoury.  TMost text is in 'Object Text'  The five shields have their text in 'Shield Object Text'  N.B. There is an SVG file embedded in the latter.  If this can be put in the box, great, if not, don't worry about it.